% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot.R
\name{construct_pearson}
\alias{construct_pearson}
\title{Compute Pearson Residuals for ARD matrix and fitted model}
\usage{
construct_pearson(y, dist = "poisson", fit = NULL, size = NULL, prob = NULL)
}
\arguments{
\item{y}{ARD matrix y}

\item{dist}{poisson or negative binomial model}

\item{fit}{(posterior) matrix means for fitted parameters}

\item{size}{matrix for negative binomial. vector of length n * k (column wise)}

\item{prob}{matrix for negative binomial. vector of length n * k (column wise)}
}
\value{
a vector (column by column) of corresponding residuals from ARD matrix
}
\description{
Compute Pearson Residuals for ARD matrix and fitted model
}
