% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_plots.R
\name{cov_plots}
\alias{cov_plots}
\title{Covariance plots}
\usage{
cov_plots(
  ard,
  model_fit,
  x_cov,
  resid_type = c("rqr", "pearson_residuals"),
  method = "lm",
  se = F
)
}
\arguments{
\item{ard}{ard matrix}

\item{model_fit}{a fitted object from [fit_mle()] or [fit_map()]}

\item{x_cov}{covariate matrix}

\item{resid_type}{the type of residuals to use}

\item{method}{the method to use}

\item{se}{whether to compute standard errors of estimates}
}
\value{
a list of ggplots, corresponding to covariance structure
}
\description{
Plots of the estimated covariance structure from a given fitted model
}
