% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_corr_perm.R
\name{degree_corr_perm}
\alias{degree_corr_perm}
\title{Permutation distribution of correlation between estimated alpha and residuals
from corresponding model}
\usage{
degree_corr_perm(alphas, resids, b = 1000, plot = TRUE)
}
\arguments{
\item{alphas}{vector of the estimated}

\item{resids}{matrix of pearson residuals for ARD matrix}

\item{b}{number of permutations}

\item{plot}{whether to show the plot (default) or not}
}
\value{
a list of test statistics
}
\description{
This function is incomplete, in progress
}
\details{
This computes the correlation between the alphas and each column 
of the residual matrix, and finds the range of these correlations,
comparing the value to those under permutations of the alpha row
}
