% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ian_Fit_models.R, R/fit_ard.R
\name{fit_stan_optim}
\alias{fit_stan_optim}
\title{Find MAP estimates of basic Poisson and Negative Binomial models using Stan optimization}
\usage{
fit_stan_optim(
  ard,
  x_cov_global = NULL,
  x_cov_local = NULL,
  family = c("poisson", "nbinomial"),
  ...
)

fit_stan_optim(
  ard,
  x_cov_global = NULL,
  x_cov_local = NULL,
  family = c("poisson", "nbinomial"),
  ...
)
}
\arguments{
\item{ard}{n_i by n_k ARD matrix}

\item{x_cov_global}{n_i by p_global covariate matrix of global covariates}

\item{x_cov_local}{n_i by p_local covariate matrix of lobal covariates}

\item{family}{Distribution to fit, either "poisson" or "nbinomial"}

\item{...}{additional arguments to the stan fit}

\item{x_cov}{n_i by p covariate matrix with columns names}

\item{cov_names}{Vector of column names to include}
}
\value{
Stan fit

Stan fit
}
\description{
Find MAP estimates of basic Poisson and Negative Binomial models using Stan optimization

Find MAP estimates of basic Poisson and Negative Binomial models using Stan optimization
}
