% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr.R
\name{get_rqr}
\alias{get_rqr}
\title{Compute Randomized Quantile Residuals for ARD Models}
\usage{
get_rqr(
  y,
  p = NULL,
  fit = NULL,
  size = NULL,
  prob = NULL,
  dist = c("binomial", "negbin", "poisson")
)
}
\arguments{
\item{y}{ard matrix}

\item{p}{if binomial distribution, success probability (single value)}

\item{fit}{if poisson distribution, rate matrix}

\item{size}{if negative binomial, size matrix}

\item{prob}{if negative binomial, size matrix}

\item{dist}{the distribution to fit, which will choose prev pars to be
specified}
}
\value{
a vector of residuals (column by column)
}
\description{
Compute Randomized Quantile Residuals for ARD Models
}
