% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hang_rootogram.R
\name{hang_rootogram_ard}
\alias{hang_rootogram_ard}
\title{Hanging Rootogram for Fitted ARD Model}
\usage{
hang_rootogram_ard(ard, model_fit, width = 0.9, x_max = NULL, by_group = FALSE)
}
\arguments{
\item{ard}{ard matrix}

\item{model_fit}{fitted model object}

\item{width}{width of bars}

\item{x_max}{the maximum x value to display}

\item{by_group}{logical; if TRUE, create separate rootograms for each column (group)}
}
\value{
a ggplot of the hanging rootogram (single plot if by_group=FALSE, combined plot if by_group=TRUE)
}
\description{
Hanging Rootogram for Fitted ARD Model
}
