% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{log_mix_uniform}
\alias{log_mix_uniform}
\title{log computed uniform quantile}
\usage{
log_mix_uniform(logFl, logFu)
}
\arguments{
\item{logFl}{log of lower value}

\item{logFu}{log of upper value}
}
\value{
log value of uniform between Flower and Fupper
}
\description{
log computed uniform quantile
}
