% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_perms.R
\name{pca_group_corr_test}
\alias{pca_group_corr_test}
\title{Title}
\usage{
pca_group_corr_test(ard, model_fit, b = 1000)
}
\arguments{
\item{ard}{ARD data}

\item{model_fit}{list containing fitted model and additional details}

\item{b}{number of replications to do}
}
\value{
list containing plots, test statistics, etc
}
\description{
Title
}
