% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{residual_correlation}
\alias{residual_correlation}
\title{Construction Residual (row/column) correlation matrix}
\usage{
residual_correlation(ard_residuals, ard, type = "column")
}
\arguments{
\item{ard_residuals}{vector of residuals}

\item{ard}{ard matrix}

\item{type}{type of correlation to use (row or column)}
}
\value{
a ggplot of the specified correlation matrix
}
\description{
Construction Residual (row/column) correlation matrix
}
