% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{residual_heatmap}
\alias{residual_heatmap}
\title{Construct heatmap of residuals}
\usage{
residual_heatmap(ard_residuals, y)
}
\arguments{
\item{ard_residuals}{a vector (column wise) of estimated residuals}

\item{y}{an ard matrix}
}
\value{
A ggplot of residual heatmap
}
\description{
Construct heatmap of residuals
}
