% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{rqr_nbinom_logs}
\alias{rqr_nbinom_logs}
\title{compute numerically stable negative binomial rqr}
\usage{
rqr_nbinom_logs(y, size, prob, eps = 1e-12)
}
\arguments{
\item{y}{observed value}

\item{size}{size parameter}

\item{prob}{prob parameter}

\item{eps}{precision parameter}
}
\value{
appropriate randomized quantile residual
}
\description{
compute numerically stable negative binomial rqr
}
