% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_perms.R
\name{tw_group_corr_test}
\alias{tw_group_corr_test}
\title{Tracy-Widom test for residual group correlation}
\usage{
tw_group_corr_test(model_fit, correction = c("none", "half"), plot = TRUE)
}
\arguments{
\item{model_fit}{fitted model object}

\item{correction}{correction constant, either "none", "half"}

\item{plot}{a logical, whether to return a ggplot density plot of TW with observed statistic}
}
\value{
a list containing test statistic, p-value, and diagnostic plots
}
\description{
Tracy-Widom test for residual group correlation
}
