% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{construct_pearson}
\alias{construct_pearson}
\title{Compute Pearson Residuals for ARD matrix and fitted model}
\usage{
construct_pearson(ard, model_fit)
}
\arguments{
\item{ard}{ARD matrix y}

\item{model_fit}{estimated model}
}
\value{
a vector (column by column) of corresponding residuals from ARD matrix
}
\description{
Compute Pearson Residuals for ARD matrix and fitted model
}
