% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{construct_pearson}
\alias{construct_pearson}
\title{Compute Pearson Residuals for ARD matrix and fitted model}
\usage{
construct_pearson(y, model_fit = NULL, family = "poisson")
}
\arguments{
\item{y}{ARD matrix y}

\item{model_fit}{estimated model}

\item{family}{poisson or negative binomial}
}
\value{
a vector (column by column) of corresponding residuals from ARD matrix
}
\description{
Compute Pearson Residuals for ARD matrix and fitted model
}
