% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{construct_rqr}
\alias{construct_rqr}
\title{Compute Randomized Quantile Residuals for ARD Models}
\usage{
construct_rqr(
  y,
  model_fit = NULL,
  family = c("binomial", "nbinomial", "poisson")
)
}
\arguments{
\item{y}{ard matrix}

\item{family}{the familyribution to fit, which will choose prev pars to be
specified}

\item{p}{if binomial familyribution, success probability (single value)}

\item{fit}{if poisson familyribution, rate matrix}

\item{size}{if negative binomial, size matrix}

\item{prob}{if negative binomial, size matrix}
}
\value{
a vector of residuals (column by column)
}
\description{
Compute Randomized Quantile Residuals for ARD Models
}
