% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{construct_rqr}
\alias{construct_rqr}
\title{Compute Randomized Quantile Residuals for ARD Models}
\usage{
construct_rqr(ard, model_fit)
}
\arguments{
\item{ard}{ard matrix}

\item{model_fit}{fitted model, along with required details}
}
\value{
a vector of residuals (column by column)
}
\description{
Compute Randomized Quantile Residuals for ARD Models
}
