% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\name{dispersion_metric}
\alias{dispersion_metric}
\title{Dispersion Metric for Fitted ARD Model}
\usage{
dispersion_metric(ard, model_fit)
}
\arguments{
\item{ard}{ard matrix}

\item{model_fit}{list of fitted model and details}
}
\value{
a ggplot of the hanging rootogram
}
\description{
Dispersion Metric for Fitted ARD Model
}
