% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ard.R
\name{fit_mle}
\alias{fit_mle}
\title{Fit basic Poisson and Negative Binomial models using glmmTMB}
\usage{
fit_mle(
  ard,
  x_cov_global = NULL,
  x_cov_local = NULL,
  family = c("poisson", "nbinomial")
)
}
\arguments{
\item{ard}{n_i by n_k ARD matrix}

\item{x_cov_global}{n_i by p_global covariate matrix of global covariates}

\item{x_cov_local}{n_i by p_local covariate matrix of local covariates}

\item{family}{Distribution to fit, either "poisson" or "nbinomial"}
}
\value{
List containing fitted model and extracted parameters
}
\description{
Fit basic Poisson and Negative Binomial models using glmmTMB
}
