% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ard.R
\name{fit_stan_optim}
\alias{fit_stan_optim}
\title{Find MAP estimates of basic Poisson and Negative Binomial models using Stan optimization}
\usage{
fit_stan_optim(
  ard,
  x_cov_global = NULL,
  x_cov_local = NULL,
  family = c("poisson", "nbinomial")
)
}
\arguments{
\item{ard}{n_i by n_k ARD matrix}

\item{x_cov_global}{n_i by p_global covariate matrix of global covariates}

\item{x_cov_local}{n_i by p_local covariate matrix of lobal covariates}

\item{family}{Distribution to fit, either "poisson" or "nbinomial"}
}
\value{
Stan fit
}
\description{
Find MAP estimates of basic Poisson and Negative Binomial models using Stan optimization
}
