% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr.R
\name{get_rqr}
\alias{get_rqr}
\title{Compute Randomized Quantile Residuals for ARD Models}
\usage{
get_rqr(y, model_fit = NULL)
}
\arguments{
\item{y}{ard matrix}

\item{model_fit}{list of details of fitted model
specified}
}
\value{
a vector of residuals (column by column)
}
\description{
Compute Randomized Quantile Residuals for ARD Models
}
