% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{get_surrogate}
\alias{get_surrogate}
\title{Compute Surrogate Residuals for ARD Models}
\usage{
get_surrogate(y, model_fit = NULL, family = c("binomial", "negbin", "poisson"))
}
\arguments{
\item{y}{the ARD matrix}

\item{model_fit}{if poisson distribution, rate matrix}

\item{family}{the distribution to fit, which will choose prev pars to be
specified}
}
\value{
a vector of residuals (column by column)
}
\description{
Compute Surrogate Residuals for ARD Models
}
