% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{get_surrogate}
\alias{get_surrogate}
\title{Compute Surrogate Residuals for ARD Models}
\usage{
get_surrogate(y, model_fit = NULL)
}
\arguments{
\item{y}{the ARD matrix}

\item{model_fit}{list containing fitted model, details}
}
\value{
a vector of residuals (column by column)
}
\description{
Compute Surrogate Residuals for ARD Models
}
