% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hang_rootogram.R
\name{hang_rootogram_ard}
\alias{hang_rootogram_ard}
\title{Hanging Rootogram for Fitted ARD Model}
\usage{
hang_rootogram_ard(
  y,
  model_fit = NULL,
  family = "poisson",
  size = NULL,
  prob = NULL,
  width = 0.9,
  x_max = NULL
)
}
\arguments{
\item{y}{ard matrix}

\item{family}{whether ard model poisson or negative binomial}

\item{size}{if negative binomial, matrix of size estimate}

\item{prob}{if negative binomial, matrix of prob estimate}

\item{width}{width of bars}

\item{x_max}{the maximum x value to display}

\item{fit}{matrix of (estimated) means of each entry if poisson ARD model}
}
\value{
a ggplot of the hanging rootogram
}
\description{
Hanging Rootogram for Fitted ARD Model
}
