% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot.R
\name{make_ard_tidy}
\alias{make_ard_tidy}
\title{Construct tibble from ARD matrix}
\usage{
make_ard_tidy(y)
}
\arguments{
\item{y}{ARD matrix}
}
\value{
a tibble of ARD, with columns for row/col index
}
\description{
Construct tibble from ARD matrix
}
