% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkscaleup-package.R
\docType{package}
\name{networkscaleup}
\alias{networkscaleup-package}
\alias{networkscaleup}
\title{The 'networkscaleup' package.}
\description{
Provides a variety of Network Scale-up Models for researchers to
  analyze Aggregated Relational Data, mostly through the use of Stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ilaga/networkscaleup}
}

}
\author{
\strong{Maintainer}: Ian Laga \email{owengward@gmail.com} (\href{https://orcid.org/0000-0002-5164-4856}{ORCID})

Authors:
\itemize{
  \item Owen G. Ward \email{oward@sfu.ca}
  \item Anna L. Smith
  \item Benjamin Vogel
  \item Jieyun Wang
  \item Le Bao \email{lebao@psu.edu}
  \item Xiaoyue Niu \email{Xiaoyue@psu.edu}
}

}
