% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot residuals against fitted values}
\usage{
plot_fitted(
  y,
  p = NULL,
  fit = NULL,
  size = NULL,
  prob = NULL,
  dist = c("binomial", "negbin", "poisson"),
  resid = c("rqr", "pearson", "surrogate")
)
}
\arguments{
\item{y}{ARD matrix (may be needed)}

\item{p}{if binomial distribution, success probability (single value)}

\item{fit}{if poisson distribution, rate matrix}

\item{size}{if negative binomial, size matrix}

\item{prob}{if negative binomial, prob matrix}

\item{dist}{the distribution to be fit}

\item{resid}{the type of residuals to be used}
}
\value{
a ggplot showing fitted values against residuals
}
\description{
Plot residuals against fitted values
}
