% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot residuals against fitted values}
\usage{
plot_fitted(
  y,
  model_fit = NULL,
  family = c("binomial", "negbin", "poisson"),
  resid = c("rqr", "pearson", "surrogate")
)
}
\arguments{
\item{y}{ARD matrix (may be needed)}

\item{model_fit}{fitted model}

\item{family}{the distribution to be fit}

\item{resid}{the type of residuals to be used}
}
\value{
a ggplot showing fitted values against residuals
}
\description{
Plot residuals against fitted values
}
