% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{residual_correlation}
\alias{residual_correlation}
\title{Construction Residual (row/column) correlation matrix}
\usage{
residual_correlation(ard_residuals, y, type = "column")
}
\arguments{
\item{ard_residuals}{}

\item{y}{ard matrix y}

\item{type}{}
}
\value{
a ggplot of the specified correlation matrix
}
\description{
Construction Residual (row/column) correlation matrix
}
