% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot.R
\name{residual_correlation}
\alias{residual_correlation}
\title{Construction Residual (row/column) correlation matrix}
\usage{
residual_correlation(ard_residuals, y, type = "column")
}
\arguments{
\item{ard_residuals}{a vector (column by column) of ard residuals}

\item{y}{ard matrix y}

\item{type}{the type of correlation to be computed (row or column)}
}
\value{
a ggplot of the correlation matrix
}
\description{
Construction Residual (row/column) correlation matrix
}
