% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot.R
\name{residual_heatmap}
\alias{residual_heatmap}
\title{Construct heatmap of ARD residuals}
\usage{
residual_heatmap(ard_residuals, y)
}
\arguments{
\item{ard_residuals}{vector (column by column) of fitted residuals}

\item{y}{ARD matrix y}
}
\value{
a heatmap of the residuals
}
\description{
Construct heatmap of ARD residuals
}
