% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{rqr_pois_logs}
\alias{rqr_pois_logs}
\title{compute numerically stable Poisson rqr}
\usage{
rqr_pois_logs(y, mu, eps = 1e-12)
}
\arguments{
\item{y}{observed value}

\item{mu}{mean value of poisson}

\item{eps}{precision parameter}
}
\value{
appropriate randomized quantile residual
}
\description{
compute numerically stable Poisson rqr
}
